requirements.txt
aiosqlite
discord.py
matplotlib
python-dotenv
msvcrt
fcntl